requirements.txt
Flask==3.0.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.35.0
sentence-transformers==3.0.1
torch>=2.1.0; platform_system != 'Darwin' or platform_machine != 'arm64'
torch==2.2.2; platform_system == 'Darwin' and platform_machine == 'arm64'
numpy==1.26.4
tiktoken==0.7.0
prometheus-client==0.20.0
